Refactor codes a bit
